storage: move many commands to `batcheval`
This spares `endTransaction` which merits its own PR due to the commit
triggers.

It also does not yet move the span declaration methods, but nothing is
blocking that.
