+++ pkg/storage/batcheval/cmd_heartbeat_txn.go
@@ -1,0 +1,64 @@
+// Copyright 2014 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package batcheval
+
+import (
+	"golang.org/x/net/context"
+
+	"github.com/cockroachdb/cockroach/pkg/keys"
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/storage/batcheval/result"
+	"github.com/cockroachdb/cockroach/pkg/storage/engine"
+	"github.com/cockroachdb/cockroach/pkg/util/hlc"
+	"github.com/pkg/errors"
+)
+
+// HeartbeatTxn updates the transaction status and heartbeat
+// timestamp after receiving transaction heartbeat messages from
+// coordinator. Returns the updated transaction.
+func HeartbeatTxn(
+	ctx context.Context, batch engine.ReadWriter, cArgs CommandArgs, resp roachpb.Response,
+) (result.Result, error) {
+	args := cArgs.Args.(*roachpb.HeartbeatTxnRequest)
+	h := cArgs.Header
+	reply := resp.(*roachpb.HeartbeatTxnResponse)
+
+	if err := VerifyTransaction(h, args); err != nil {
+		return result.Result{}, err
+	}
+
+	key := keys.TransactionKey(h.Txn.Key, h.Txn.ID)
+
+	var txn roachpb.Transaction
+	if ok, err := engine.MVCCGetProto(ctx, batch, key, hlc.Timestamp{}, true, nil, &txn); err != nil {
+		return result.Result{}, err
+	} else if !ok {
+		// If no existing transaction record was found, skip heartbeat.
+		// This could mean the heartbeat is a delayed relic or it could
+		// mean that the BeginTransaction call was delayed. In either
+		// case, there's no reason to persist a new transaction record.
+		return result.Result{}, errors.Errorf("heartbeat for transaction %s failed; record not present", h.Txn)
+	}
+
+	if txn.Status == roachpb.PENDING {
+		txn.LastHeartbeat.Forward(args.Now)
+		if err := engine.MVCCPutProto(ctx, batch, cArgs.Stats, key, hlc.Timestamp{}, nil, &txn); err != nil {
+			return result.Result{}, err
+		}
+	}
+
+	reply.Txn = &txn
+	return result.Result{}, nil
+}
