+++ pkg/storage/batcheval/cmd_lease_transfer.go
@@ -1,0 +1,44 @@
+// Copyright 2014 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package batcheval
+
+import (
+	"golang.org/x/net/context"
+
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/storage/batcheval/result"
+	"github.com/cockroachdb/cockroach/pkg/storage/engine"
+	"github.com/cockroachdb/cockroach/pkg/util/log"
+)
+
+// TransferLease sets the lease holder for the range.
+// Unlike with RequestLease(), the new lease is allowed to overlap the old one,
+// the contract being that the transfer must have been initiated by the (soon
+// ex-) lease holder which must have dropped all of its lease holder powers
+// before proposing.
+func TransferLease(
+	ctx context.Context, batch engine.ReadWriter, cArgs CommandArgs, resp roachpb.Response,
+) (result.Result, error) {
+	args := cArgs.Args.(*roachpb.TransferLeaseRequest)
+
+	// When returning an error from this method, must always return
+	// a newFailedLeaseTrigger() to satisfy stats.
+	prevLease, _ := cArgs.EvalCtx.GetLease()
+	if log.V(2) {
+		log.Infof(ctx, "lease transfer: prev lease: %+v, new lease: %+v", prevLease, args.Lease)
+	}
+	return evalNewLease(ctx, cArgs.EvalCtx, batch, cArgs.Stats,
+		args.Lease, prevLease, false /* isExtension */, true /* isTransfer */)
+}
